third_party/blink/web_tests/platform/generic/fast/xpath/4XPath/Borrowed/cz_20030217-expected.txt

PASS nodeset.snapshotLength is 1
PASS nodeset.snapshotLength is 0
PASS nodeset.snapshotLength is 1
PASS nodeset.snapshotLength is 0
PASS successfullyParsed is true

TEST COMPLETE
